Internal dg_geom_plot functionality (#24)
Previously, a system call was being made to dg_geom_plot.py, which is a command line function contained in the dgtools library. We are trying to move away from this library, so the functionality is ported here.

Largely this was just copying over methods and imposing some structure on them.

This PR removes any existing dgtools dependency from the asp_plot package.