Improve title and filename for report (#29)
This PR allows a user to supply title for the report, and extends the functionality of the current filename for the report. If no title or filename is supplied, then the default is to use the name of the directory where the processed files are stored (which usually contains some identifying information about the scenes and their dates).

Also snuck in osr.UseExceptions() to silence a deprecation warning.