Gracefully handle missing files (#31)
The glob.glob functions used throughout the codebase to gather files are ported to one flexible function that warns a user when a file is not found. In most cases (except critical ones) the plotting proceeds and missing data is handled by displaying the plot with a text notice that one or more files is missing. At that point, a user will need to pay attention to warning messages they received to track down and generate the missing files. These are likely edge cases, as most users will not be deleting the automatically generated ASP file outputs.