Trim linescan csm plot by acquisition and plot line numbers (#57)
Previously, the csm_camera_plot was just plotting all of the positions and orientations, even those outside the image area. Or, optionally, a user could pass in a "trim" parameter that would try to shave off the beginning and end of the positions. However, this was difficult to use and not based on the physical collection of the image as it relates to the CSM model.

Now, the "trim" parameter actually gets the real first and last lines of the image from the CSM model and trims by those. In addition, the approximate scene line number is plotted on the x-axis.