Run, report, and translate with pc_align for multi-processed altimetry (#68)
This PR lays some foundational groundwork built on top of the multi-processed ICESat-2 retrieval from PR #59. 

Namely, the `pc_align` functionality has been moved to an `asp_plot.alignment` module in the `Alignment` class. This `Alignment` class is then instantiated from within the `Altimetry` class, allowing the multi-processed ICESat-2 data to spawn several pc_align runs. 

The results of these pc_align runs on the various temporally processed ICESat-2 points are reported to the user, and, if the translation is above some threshold relative to the DEM's GSD, the alignment is applied. However, when the alignment is applied, the user is notified of any gross disagreements between the translation results for the different time periods of ICESat-2 points used in the alignment, which may indicate unreliable alignment and warrant caution.

It is left to the user to decide how to proceed at that point.