Remove high_confidence from SlideRule retrieval
For a discussion of this, see: https://github.com/SlideRuleEarth/sliderule/issues/448.

The main point is that "high_confidence" points are not necessarily a given surface type. Instead, we want to focus on just getting the ground, canopy, and top of canopy, which we already do.

This commit removes all traces of "high_confidence" points from the codebase.

In addition, this commit speeds up the altimetry tests by skipping the actual external retrieval call. This means the `request_atl06sr_multi_processing` function is not tested, but that is mostly a wrapper for sliderule anyway.

Tests are passing locally. Failing tests on CI may have to do with the CI setup on GitHub.