Remove high_confidence from SlideRule retrieval (#72)
This commit removes all traces of "high_confidence" points from the codebase. For a discussion of this, see: https://github.com/SlideRuleEarth/sliderule/issues/448.

The main point is that "high_confidence" points are not necessarily a given surface type. Instead, we want to focus on just getting the ground, canopy, and top of canopy, which we already do.

In addition, this commit speeds up the altimetry tests by skipping the actual external retrieval call. This means the `request_atl06sr_multi_processing` function is not tested, but that is mostly a wrapper for sliderule anyway.

Also, I fixed CI tests here, by dropping use of mamba in favor of conda, which seems to have fixed things.