Increase flexibility of stereo DEM filename search (#77)
Previously, the StereoPlotter tool required a GSD specified in the DEM filename in the stereo directory. This is non-standard for ASP output.

Now the instantiation of the class if flexible allowing for:

- An optional `dem_gsd` argument, which will search for a specified GSD in the `*-DEM` file in the stereo directory
- An optional `dem_fn` argument, which will override the GSD and default search and use the custom filename provided by the user.

These arguments were also added to the CLI and specs were updated to check for the different use cases.

---------

Co-authored-by: Ben Purinton <ben.purinton@gmail.com>