Geometry plotting refactoring and cleanup (#78)
The existing metadata parsing and scene geometry plotting was previously ported as-is from dgtools. This PR refactors the code to be more flexible for varied XML file inputs. The focus is still on WV stereopairs, but the improvements should be a good base for increasing flexibility for other satellites and more than two scene inputs.

BIE and Assymetry angles were also added to the output geometry plot to resolve #19

---------

Co-authored-by: Ben Purinton <ben.purinton@gmail.com>