Partial ICESat cleanup and add ability to run without internet (#81)
This commit allows asp_plot to run offline by setting `--add_basemap False --plot_icesat False`.

The SlideRule parameters are also refactored and updated to better select "all", "ground", "canopy", and "top_of_canopy" measurements.

Finally, the SlideRule request parameters are now stored with the parquet files, instead of in the filename. If a request is made with different parameters, the existing file is checked and regenerated if the request has different parameters than the original.

A number of TODOs remain in `altimetry.py`, but these should be handled along with remaining updates on issues #62 and #40 on dedicated, smaller pull requests.

---------

Co-authored-by: Ben Purinton <ben.purinton@gmail.com>